requirement.txt
tweepy==4.14.0
praw==7.7.1
requests==2.31.0
vaderSentiment==3.3.2
transformers==4.41.1
torch>=2.0.0
streamlit==1.35.0
plotly==5.22.0
pandas==2.2.2
beautifulsoup4==4.12.3
lxml==5.2.1
selenium==4.21.0
python-dotenv==1.0.1
google-cloud-storage==2.16.0
google-auth==2.29.0
wordcloud==1.9.3
matplotlib==3.9.0

